+++ src/main/java/co/orange/ddanzi/domain/user/InterestProduct.java
@@ -1,6 +1,6 @@
-package co.orange.ddanzi.domain.member;
+package co.orange.ddanzi.domain.user;
 
-import co.orange.ddanzi.domain.member.pk.InterestProductId;
+import co.orange.ddanzi.domain.user.pk.InterestProductId;
 import jakarta.persistence.Column;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
