+++ src/main/java/co/orange/ddanzi/domain/user/UserAgreement.java
@@ -1,7 +1,7 @@
-package co.orange.ddanzi.domain.member;
+package co.orange.ddanzi.domain.user;
 
 import co.orange.ddanzi.common.domain.BaseTimeEntity;
-import co.orange.ddanzi.domain.member.pk.UserAgreementId;
+import co.orange.ddanzi.domain.user.pk.UserAgreementId;
 import jakarta.persistence.Column;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
